+++ src/js/kit.js
@@ -12,5 +12,5 @@
 // @codekit-prepend "components/autoheight-carousel.js"
 // @codekit-prepend "components/autoheight-gallery.js"
 // @codekit-prepend "components/offcanvas.js"
-// codekit-prepend "components/sticky.js"
+// @codekit-prepend "components/sticky.js"
 // @codekit-prepend "components/equalizer.js"